+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 `WhereAmI` adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0](https://github.com/lypiut/WhereAmI/releases/tag/2.1.0)
+
+Rewrite of the library in order to be more swift friendly and easy to use.
+
+#### Added
+- `LocationProfile` protocol allows you to create your custom location profile
+- The `WhereAmI` and `whatIsThisPlace` methods have one closure for the response handling and the `locationRefusedHandler` has been removed. Responses are returned via an enum.
+
 ## [2.1.0](https://github.com/lypiut/WhereAmI/releases/tag/2.1.0)
 
 #### Added